requirements.txt
django
django-simpleui
djangorestframework
django-cors-headers
drf-yasg
djangorestframework-simplejwt
pillow